1.18.2 backport because I was bored today

+++ src/main/java/io/github/bumblesoftware/fastload/config/screen/BuildingTerrainScreen.java
@@ -38,7 +38,7 @@ public class BuildingTerrainScreen extends Screen {
     public void render(MatrixStack matrices, int mouseX, int mouseY, float delta) {
         this.renderBackgroundTexture(0);
         final String loadedChunksString = getLoadedChunkCount() + "/"  + FLMath.getPreRenderArea();
-        final String builtChunksString = getBuiltChunkCount() + "/"  + FLMath.getPreRenderArea() * client.options.fov / 360;
+        final String builtChunksString = getBuiltChunkCount() + "/"  + (int) (FLMath.getPreRenderArea() * client.options.fov / 360);
         if (PREPARED_PROGRESS_STORAGE < getLoadedChunkCount()) {
             FastLoad.LOGGER.info("World Chunk Sending: " + loadedChunksString);
         }
